backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
pydantic-settings>=2.2.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
cryptography>=42.0.0
requests-oauthlib>=1.3.1
python-dotenv>=1.0.0
openai>=1.0.0
